# docs-site/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Greenlight — Multi-Platform Prompt & Channel Gateway",
    "version": "0.1.0",
    "description": "Self-hosted HTTP gateway for human-in-the-loop AI agents across Telegram, Slack, Teams, Discord, Google Chat, WhatsApp, and Messenger.\n\n**Path verbs:** `POST …/new` creates a durable resource (channel or prompt). `POST …/send` delivers chat text on a MESSAGE channel only.",
    "license": {
      "name": "BUSL-1.1",
      "url": "https://spdx.org/licenses/BUSL-1.1.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8100",
      "description": "Local development"
    },
    {
      "url": "https://{host}",
      "description": "Self-hosted production (PUBLIC_WEBHOOK_URL origin)",
      "variables": {
        "host": {
          "default": "api.example.com",
          "description": "Your Greenlight host (no scheme)"
        }
      }
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Database-backed API key with scoped permissions. Bootstrap from GREENLIGHT_API_KEY env on first boot when api_keys is empty."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      },
...
      "RetentionSettings": {
        "type": "object",
        "properties": {
          "prompts_retention_enabled": {
            "type": "boolean"
          },
          "prompts_retention_days": {
            "type": "integer",
            "minimum": 1,
            "maximum": 3650
          },
          "messages_inbound_retention_enabled": {
            "type": "boolean"
          },
          "messages_outbound_retention_enabled": {
            "type": "boolean"
          },
          "messages_inbound_retention_days": {
            "type": "integer",
            "minimum": 1,
            "maximum": 3650
          },
          "messages_outbound_retention_days": {
            "type": "integer",
            "minimum": 1,
            "maximum": 3650
          },
          "messages_inbound_zero_retention": {
            "type": "boolean"
          },
          "messages_outbound_zero_retention": {
            "type": "boolean"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get data retention settings",
        "description": "Requires `settings:read` scope (or `admin`).",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Retention settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionSettings"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Missing required scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Update data retention settings",
        "description": "Requires `settings:write` scope (or `admin`).",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionSettings"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Missing required scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}